+++ examples/examples_test.go
@@ -32,9 +32,16 @@ func TestExamples(t *testing.T) {
 		},
 	})
 
+	baseDotNet := base.With(integration.ProgramTestOptions{
+		Dependencies: []string{
+			"Pulumi.Terraform",
+		},
+	})
+
 	shortTests := []integration.ProgramTestOptions{
 		baseJS.With(integration.ProgramTestOptions{Dir: path.Join(cwd, "localstate-nodejs")}),
 		basePython.With(integration.ProgramTestOptions{Dir: path.Join(cwd, "localstate-python")}),
+		baseDotNet.With(integration.ProgramTestOptions{Dir: path.Join(cwd, "localstate-dotnet")}),
 	}
 
 	longTests := []integration.ProgramTestOptions{}