+++ pkg/tfgen/generate_nodejs.go
@@ -1051,8 +1051,15 @@ func (g *nodeJSGenerator) emitResourceFunc(mod *module, fun *resourceFunc, neste
 	} else {
 		retty = fun.retst.name
 	}
-	w.Writefmtln("export function %s(%sopts?: pulumi.InvokeOptions): Promise<%s> & %s {",
-		fun.name, argsig, retty, retty)
+
+	if g.info.JavaScript.AsyncDataSources {
+		w.Writefmtln("export function %s(%sopts?: pulumi.InvokeOptions): Promise<%s> {",
+			fun.name, argsig, retty)
+	} else {
+		w.Writefmtln("export function %s(%sopts?: pulumi.InvokeOptions): Promise<%s> & %s {",
+			fun.name, argsig, retty, retty)
+	}
+
 	if fun.info.DeprecationMessage != "" {
 		w.Writefmtln("    pulumi.log.warn(\"%s is deprecated: %s\")", fun.name, fun.info.DeprecationMessage)
 	}
@@ -1071,16 +1078,24 @@ func (g *nodeJSGenerator) emitResourceFunc(mod *module, fun *resourceFunc, neste
 	w.Writefmtln("        opts.version = utilities.getVersion();")
 	w.Writefmtln("    }")
 
-	// Now simply invoke the runtime function with the arguments, returning the results.
-	w.Writefmtln("    const promise: Promise<%s> = pulumi.runtime.invoke(\"%s\", {", retty, fun.info.Tok)
+	if g.info.JavaScript.AsyncDataSources {
+		// Now simply invoke the runtime function with the arguments, returning the results.
+		w.Writefmtln("    return pulumi.runtime.invoke(\"%s\", {", fun.info.Tok)
+	} else {
+		w.Writefmtln("    const promise: Promise<%s> = pulumi.runtime.invoke(\"%s\", {", retty, fun.info.Tok)
+	}
+
 	for _, arg := range fun.args {
 		// Pass the argument to the invocation.
 		w.Writefmtln("        \"%[1]s\": args.%[1]s,", arg.name)
 	}
 	w.Writefmtln("    }, opts);")
 	w.Writefmtln("")
-	w.Writefmtln("    return pulumi.utils.liftProperties(promise, opts);")
-	w.Writefmtln("}")
+
+	if !g.info.JavaScript.AsyncDataSources {
+		w.Writefmtln("")
+		w.Writefmtln("    return pulumi.utils.liftProperties(promise, opts);")
+	}
 
 	// If there are argument and/or return types, emit them.
 	if fun.argst != nil {