+++ scripts/publish-plugin.sh
@@ -16,7 +16,7 @@ if [ "$(go env GOOS)" = "windows" ]; then
     BIN_SUFFIX=".exe"
 fi
 
-go build \
+cd "${ROOT}/provider" && go build \
    -ldflags "-X github.com/pulumi/pulumi-terraform/provider/cmd/pulumi-resource-terraform/main.Version=${VERSION}" \
    -o "${WORK_PATH}/pulumi-resource-terraform${BIN_SUFFIX}" \
    "${ROOT}/provider/cmd/pulumi-resource-terraform"