+++ CHANGELOG.md
@@ -2,6 +2,11 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
+_(none)_
+
+---
+
+## v2.0.0 (2020-05-02)
 * `RemoteStateReference` resources can now read states created with Terraform v0.12.24 and below
 * Fix a bug with `RemoteStateReference`'s remote state backend on Python, where the workspace name was not  getting configured correctly (issue [#524](https://github.com/pulumi/pulumi-terraform/issues/524)).
 * Added support for .NET with the `Pulumi.Terraform` NuGet package